move weight trader tests into separate mod
