+++ system-parachains/asset-hubs/asset-hub-kusama/tests/weight_trader.rs
@@ -32,6 +32,13 @@ use frame_support::{
 	},
 	weights::{Weight, WeightToFee as WeightToFeeT},
 };
+
+use asset_hub_kusama_runtime::xcm_config::AssetFeeAsExistentialDepositMultiplierFeeCharger;
+use asset_hub_kusama_runtime::xcm_config::TrustBackedAssetsPalletLocation;
+use asset_hub_kusama_runtime::ExistentialDeposit;
+use cumulus_primitives_utility::ChargeWeightInFungibles;
+use frame_support::assert_noop;
+
 use parachains_common::{AccountId, AuraId};
 use sp_runtime::traits::MaybeEquivalence;
 use system_parachains_constants::kusama::{currency::*, fee::WeightToFee};
@@ -42,6 +49,347 @@ const ALICE: [u8; 32] = [1u8; 32];
 const SOME_ASSET_ADMIN: [u8; 32] = [5u8; 32];
 
 type RuntimeHelper = asset_test_utils::RuntimeHelper<Runtime, AllPalletsWithoutSystem>;
+
+type AssetIdForTrustBackedAssetsConvertLatest =
+	assets_common::AssetIdForTrustBackedAssetsConvertLatest<TrustBackedAssetsPalletLocation>;
+
+#[test]
+fn test_asset_xcm_trader() {
+	ExtBuilder::<Runtime>::default()
+		.with_collators(vec![AccountId::from(ALICE)])
+		.with_session_keys(vec![(
+			AccountId::from(ALICE),
+			AccountId::from(ALICE),
+			SessionKeys { aura: AuraId::from(sp_core::sr25519::Public::from_raw(ALICE)) },
+		)])
+		.build()
+		.execute_with(|| {
+			// We need root origin to create a sufficient asset
+			let minimum_asset_balance = 3333333_u128;
+			let local_asset_id = 1;
+			assert_ok!(Assets::force_create(
+				RuntimeHelper::root_origin(),
+				local_asset_id.into(),
+				AccountId::from(ALICE).into(),
+				true,
+				minimum_asset_balance
+			));
+
+			// We first mint enough asset for the account to exist for assets
+			assert_ok!(Assets::mint(
+				RuntimeHelper::origin_of(AccountId::from(ALICE)),
+				local_asset_id.into(),
+				AccountId::from(ALICE).into(),
+				minimum_asset_balance
+			));
+
+			// get asset id as location
+			let asset_location =
+				AssetIdForTrustBackedAssetsConvertLatest::convert_back(&local_asset_id).unwrap();
+
+			// Set Alice as block author, who will receive fees
+			RuntimeHelper::run_to_block(2, AccountId::from(ALICE));
+
+			// We are going to buy 4e9 weight
+			let bought = Weight::from_parts(4_000_000_000u64, 0);
+
+			// Lets calculate amount needed
+			let asset_amount_needed =
+				AssetFeeAsExistentialDepositMultiplierFeeCharger::charge_weight_in_fungibles(
+					local_asset_id,
+					bought,
+				)
+				.expect("failed to compute");
+
+			// Lets pay with: asset_amount_needed + asset_amount_extra
+			let asset_amount_extra = 100_u128;
+			let asset: Asset =
+				(asset_location.clone(), asset_amount_needed + asset_amount_extra).into();
+
+			let mut trader = <XcmConfig as xcm_executor::Config>::Trader::new();
+			let ctx = XcmContext { origin: None, message_id: XcmHash::default(), topic: None };
+
+			// Lets buy_weight and make sure buy_weight does not return an error
+			let unused_assets = trader.buy_weight(bought, asset.into(), &ctx).expect("Expected Ok");
+			// Check whether a correct amount of unused assets is returned
+			assert_ok!(unused_assets.ensure_contains(&(asset_location, asset_amount_extra).into()));
+
+			// Drop trader
+			drop(trader);
+
+			// Make sure author(Alice) has received the amount
+			assert_eq!(
+				Assets::balance(local_asset_id, AccountId::from(ALICE)),
+				minimum_asset_balance + asset_amount_needed
+			);
+
+			// We also need to ensure the total supply increased
+			assert_eq!(
+				Assets::total_supply(local_asset_id),
+				minimum_asset_balance + asset_amount_needed
+			);
+		});
+}
+
+#[test]
+fn test_asset_xcm_trader_with_refund() {
+	ExtBuilder::<Runtime>::default()
+		.with_collators(vec![AccountId::from(ALICE)])
+		.with_session_keys(vec![(
+			AccountId::from(ALICE),
+			AccountId::from(ALICE),
+			SessionKeys { aura: AuraId::from(sp_core::sr25519::Public::from_raw(ALICE)) },
+		)])
+		.build()
+		.execute_with(|| {
+			// We need root origin to create a sufficient asset
+			// We set existential deposit to be identical to the one for Balances first
+			assert_ok!(Assets::force_create(
+				RuntimeHelper::root_origin(),
+				1.into(),
+				AccountId::from(ALICE).into(),
+				true,
+				ExistentialDeposit::get()
+			));
+
+			// We first mint enough asset for the account to exist for assets
+			assert_ok!(Assets::mint(
+				RuntimeHelper::origin_of(AccountId::from(ALICE)),
+				1.into(),
+				AccountId::from(ALICE).into(),
+				ExistentialDeposit::get()
+			));
+
+			let mut trader = <XcmConfig as xcm_executor::Config>::Trader::new();
+			let ctx = XcmContext { origin: None, message_id: XcmHash::default(), topic: None };
+
+			// Set Alice as block author, who will receive fees
+			RuntimeHelper::run_to_block(2, AccountId::from(ALICE));
+
+			// We are going to buy 4e9 weight
+			let bought = Weight::from_parts(4_000_000_000u64, 0);
+
+			let asset_location =
+				AssetIdForTrustBackedAssetsConvertLatest::convert_back(&1).unwrap();
+
+			// lets calculate amount needed
+			let amount_bought = WeightToFee::weight_to_fee(&bought);
+
+			let asset: Asset = (asset_location.clone(), amount_bought).into();
+
+			// Make sure buy_weight does not return an error
+			assert_ok!(trader.buy_weight(bought, asset.clone().into(), &ctx));
+
+			// Make sure again buy_weight does return an error
+			// This assert relies on the fact, that we use `TakeFirstAssetTrader` in `WeightTrader`
+			// tuple chain, which cannot be called twice
+			assert_noop!(trader.buy_weight(bought, asset.into(), &ctx), XcmError::TooExpensive);
+
+			// We actually use half of the weight
+			let weight_used = bought / 2;
+
+			// Make sure refurnd works.
+			let amount_refunded = WeightToFee::weight_to_fee(&(bought - weight_used));
+
+			assert_eq!(
+				trader.refund_weight(bought - weight_used, &ctx),
+				Some((asset_location, amount_refunded).into())
+			);
+
+			// Drop trader
+			drop(trader);
+
+			// We only should have paid for half of the bought weight
+			let fees_paid = WeightToFee::weight_to_fee(&weight_used);
+
+			assert_eq!(
+				Assets::balance(1, AccountId::from(ALICE)),
+				ExistentialDeposit::get() + fees_paid
+			);
+
+			// We also need to ensure the total supply increased
+			assert_eq!(Assets::total_supply(1), ExistentialDeposit::get() + fees_paid);
+		});
+}
+
+#[test]
+fn test_asset_xcm_trader_refund_not_possible_since_amount_less_than_ed() {
+	ExtBuilder::<Runtime>::default()
+		.with_collators(vec![AccountId::from(ALICE)])
+		.with_session_keys(vec![(
+			AccountId::from(ALICE),
+			AccountId::from(ALICE),
+			SessionKeys { aura: AuraId::from(sp_core::sr25519::Public::from_raw(ALICE)) },
+		)])
+		.build()
+		.execute_with(|| {
+			// We need root origin to create a sufficient asset
+			// We set existential deposit to be identical to the one for Balances first
+			assert_ok!(Assets::force_create(
+				RuntimeHelper::root_origin(),
+				1.into(),
+				AccountId::from(ALICE).into(),
+				true,
+				ExistentialDeposit::get()
+			));
+
+			let mut trader = <XcmConfig as xcm_executor::Config>::Trader::new();
+			let ctx = XcmContext { origin: None, message_id: XcmHash::default(), topic: None };
+
+			// Set Alice as block author, who will receive fees
+			RuntimeHelper::run_to_block(2, AccountId::from(ALICE));
+
+			// We are going to buy small amount
+			let bought = Weight::from_parts(50_000_000u64, 0);
+
+			let asset_location =
+				AssetIdForTrustBackedAssetsConvertLatest::convert_back(&1).unwrap();
+
+			let amount_bought = WeightToFee::weight_to_fee(&bought);
+
+			assert!(
+				amount_bought < ExistentialDeposit::get(),
+				"we are testing what happens when the amount does not exceed ED"
+			);
+
+			let asset: Asset = (asset_location, amount_bought).into();
+
+			// Buy weight should return an error
+			assert_noop!(trader.buy_weight(bought, asset.into(), &ctx), XcmError::TooExpensive);
+
+			// not credited since the ED is higher than this value
+			assert_eq!(Assets::balance(1, AccountId::from(ALICE)), 0);
+
+			// We also need to ensure the total supply did not increase
+			assert_eq!(Assets::total_supply(1), 0);
+		});
+}
+
+#[test]
+fn test_that_buying_ed_refund_does_not_refund() {
+	ExtBuilder::<Runtime>::default()
+		.with_collators(vec![AccountId::from(ALICE)])
+		.with_session_keys(vec![(
+			AccountId::from(ALICE),
+			AccountId::from(ALICE),
+			SessionKeys { aura: AuraId::from(sp_core::sr25519::Public::from_raw(ALICE)) },
+		)])
+		.build()
+		.execute_with(|| {
+			// We need root origin to create a sufficient asset
+			// We set existential deposit to be identical to the one for Balances first
+			assert_ok!(Assets::force_create(
+				RuntimeHelper::root_origin(),
+				1.into(),
+				AccountId::from(ALICE).into(),
+				true,
+				ExistentialDeposit::get()
+			));
+
+			let mut trader = <XcmConfig as xcm_executor::Config>::Trader::new();
+			let ctx = XcmContext { origin: None, message_id: XcmHash::default(), topic: None };
+
+			// Set Alice as block author, who will receive fees
+			RuntimeHelper::run_to_block(2, AccountId::from(ALICE));
+
+			// We are gonna buy ED
+			let bought = Weight::from_parts(ExistentialDeposit::get().try_into().unwrap(), 0);
+
+			let asset_location =
+				AssetIdForTrustBackedAssetsConvertLatest::convert_back(&1).unwrap();
+
+			let amount_bought = WeightToFee::weight_to_fee(&bought);
+
+			assert!(
+				amount_bought < ExistentialDeposit::get(),
+				"we are testing what happens when the amount does not exceed ED"
+			);
+
+			// We know we will have to buy at least ED, so lets make sure first it will
+			// fail with a payment of less than ED
+			let asset: Asset = (asset_location.clone(), amount_bought).into();
+			assert_noop!(trader.buy_weight(bought, asset.into(), &ctx), XcmError::TooExpensive);
+
+			// Now lets buy ED at least
+			let asset: Asset = (asset_location, ExistentialDeposit::get()).into();
+
+			// Buy weight should work
+			assert_ok!(trader.buy_weight(bought, asset.into(), &ctx));
+
+			// Should return None. We have a specific check making sure we dont go below ED for
+			// drop payment
+			assert_eq!(trader.refund_weight(bought, &ctx), None);
+
+			// Drop trader
+			drop(trader);
+
+			// Make sure author(Alice) has received the amount
+			assert_eq!(Assets::balance(1, AccountId::from(ALICE)), ExistentialDeposit::get());
+
+			// We also need to ensure the total supply increased
+			assert_eq!(Assets::total_supply(1), ExistentialDeposit::get());
+		});
+}
+
+#[test]
+fn test_asset_xcm_trader_not_possible_for_non_sufficient_assets() {
+	ExtBuilder::<Runtime>::default()
+		.with_collators(vec![AccountId::from(ALICE)])
+		.with_session_keys(vec![(
+			AccountId::from(ALICE),
+			AccountId::from(ALICE),
+			SessionKeys { aura: AuraId::from(sp_core::sr25519::Public::from_raw(ALICE)) },
+		)])
+		.build()
+		.execute_with(|| {
+			// Create a non-sufficient asset with specific existential deposit
+			let minimum_asset_balance = 1_000_000_u128;
+			assert_ok!(Assets::force_create(
+				RuntimeHelper::root_origin(),
+				1.into(),
+				AccountId::from(ALICE).into(),
+				false,
+				minimum_asset_balance
+			));
+
+			// We first mint enough asset for the account to exist for assets
+			assert_ok!(Assets::mint(
+				RuntimeHelper::origin_of(AccountId::from(ALICE)),
+				1.into(),
+				AccountId::from(ALICE).into(),
+				minimum_asset_balance
+			));
+
+			let mut trader = <XcmConfig as xcm_executor::Config>::Trader::new();
+			let ctx = XcmContext { origin: None, message_id: XcmHash::default(), topic: None };
+
+			// Set Alice as block author, who will receive fees
+			RuntimeHelper::run_to_block(2, AccountId::from(ALICE));
+
+			// We are going to buy 4e9 weight
+			let bought = Weight::from_parts(4_000_000_000u64, 0);
+
+			// lets calculate amount needed
+			let asset_amount_needed = WeightToFee::weight_to_fee(&bought);
+
+			let asset_location =
+				AssetIdForTrustBackedAssetsConvertLatest::convert_back(&1).unwrap();
+
+			let asset: Asset = (asset_location, asset_amount_needed).into();
+
+			// Make sure again buy_weight does return an error
+			assert_noop!(trader.buy_weight(bought, asset.into(), &ctx), XcmError::TooExpensive);
+
+			// Drop trader
+			drop(trader);
+
+			// Make sure author(Alice) has NOT received the amount
+			assert_eq!(Assets::balance(1, AccountId::from(ALICE)), minimum_asset_balance);
+
+			// We also need to ensure the total supply NOT increased
+			assert_eq!(Assets::total_supply(1), minimum_asset_balance);
+		});
+}
 
 #[test]
 fn test_buy_and_refund_weight_with_native() {
